extract_sources works with Windows paths
Prior to this backtrace lines were simply split by a single colon.
Unfortunately that is also the drive letter delimiter in Windows paths
which resulted in a lot of empty source fragments of "C:0". Now the
trace line is split by the first colon followed by some digits, which
works for both Windows and Unix path styles.