Warn on double render instead of disallowing it
Getting the exception while just putting a Kernel#console after
Kernel#console while tracking the error may get you out of the debug
flow state. I like this new behaviour, but it has drawbacks. The problem
with it is that you may not get the binding that you wanted, but the
binding of the last evaluated Kernel#console. I'll test a release with
it.